+++ src/components/2018/page-pwa-with-ionic-angular-wordpress-api-2/page-pwa-with-ionic-angular-wordpress-api-2.tsx
@@ -17,14 +17,16 @@ export class PagePwaWithIonicAngularWordpressApi2 {
   // header for this individual item by id...
   header: any;
 
+  id:string;
+
   async componentWillLoad() {
     if (isLocal()) {
       console.log('> PagePwaWithIonicAngularWordpressApi2.componentWillLoad');
     }
     // this.data = await BlogData.load();
     // Get the id from the URL path (slug)
-    let id = document.location.pathname.substr(1);
-    this.header = BlogData.getPostHeaderById(id);
+    this.id = document.location.pathname.substr(1);
+    this.header = BlogData.getPostHeaderById(this.id);
 
     // set document title for browser / tab / bookmark
     document.title = this.header.title + ' | ' + SITENAME;
@@ -246,6 +248,8 @@ Home
 
               <p><ion-button color="primary" routerDirection="back" href="/pwa-with-ionic-angular-wordpress-api-1">&lt;&lt; Previous: Part 1</ion-button> <ion-button color="primary" routerDirection="forward" href="/pwa-with-ionic-angular-wordpress-api-3">Next: Part 3 &gt;&gt;</ion-button></p>
               
+              <gls-disqus url={'https://codyburleson.com/' + this.id} identifier={this.id} title={this.header.title} category="" language="" />
+
             </ion-col>
             <ion-col size-xs="12" size-sm="12" size-md="4" size-lg="4" size-xl="5">
               <gls-adsense-ad />