+++ changelog.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED 002] - 2020-01-28]
+
+- Testing Ionic 5.0.0-rc.0
+
 ## [UNRELEASED 001] - 2020-01-25]
 
 - New attempt at DISQUS component (STAGING)